Fixes broken reporting id field lookup.
Seems like cartodb doesn't support uppercase identifiers in certain
fields. When we lookup the `reportingId` field, cartodb always returns
`reportingid` (lowercase i), and thus the app never finds that field,
defaulting to using the `cartodb_id` field instead on report requests.

This commit changes the field lookup to be `reporting_id` instead,
which doesn't have this problem and it also makes it more consistent
with the rest of the identifiers we have on carto.